GitSync stops depending on the AI engine — the dependency inverts
`MeshWeaver.GitSync` referenced `MeshWeaver.AI`, and through it so did `MeshWeaver.PluginCatalog`
(transitively — which is why no grep of direct ProjectReferences ever showed it). A module cannot be
injected into a platform that compiles against it, so this edge had to point the other way. Step B2
of #2276.

Five files move into `MeshWeaver.AI`: `AiContentDiskWriter`, `AiContentSyncArea`,
`AiContentSyncMenuProvider`, `HubAiContentSyncExtensions` and the drafter. None of them imported a
GitSync namespace — they were AI features filed under the wrong project. `MeshWeaver.AI` now
references `MeshWeaver.GitSync` for the repo client, which is the allowed direction.

**The PR drafter splits at an interface, and gains a platform default.** `IPullRequestDraftService`
and the `PullRequestDraft` record stay in GitSync; the agent-backed implementation goes to
`MeshWeaver.AI`. GitSync registers `PlaceholderPullRequestDraftService` with `TryAddSingleton` and
`AddAI()` registers the real one with a plain `AddSingleton` — order-independent, module always
wins (the `IEmailSender` shape). So PR creation no longer DEPENDS on AI: without it you get a
deterministic title and a body that asks you to edit it, which is what the call site's `.Catch`
already produced when no model was configured.

The registrations move with the code: the disk writer, the "Sync to repo" node-menu provider and
its area now ride `AddAI()`.

`MeshWeaver.PluginCatalog.Test` gains an EXPLICIT reference to `MeshWeaver.AI` — it installs
Agent/Skill packages and asserts on the pickers, a real AI dependency that had been arriving through
GitSync. `AiContentSyncBackTest` gains `.AddAI()` for the same reason, and its failure without it is
the proof the seam is real.

Verified with the full solution build under CI's flags and the GitSync suite: 182 tests, 0 failed.
